Skip writing the response to cache if the graphql response returns an error
Previously if there was data in the cache but the request started
returning an error, the new response would be written to the cache
with an empty `data` object, essentially wiping out the good data.
The standard Apollo behavior is to skip writing the data to the
cache on an error, leaving the good data there to be rendered.
This commit preserves that same behavior for optimistic fetch.